+++ Snake/HorizontalLine.cs
@@ -6,27 +6,16 @@ using System.Threading.Tasks;
 
 namespace Snake
 {
-    class VerticalLine
+    class HorizontalLine : Figure
     {
-        List<Point> pList;
-
-        public VerticalLine(int yUp, int yDown, int x, char sym)
+        public HorizontalLine(int xLeft, int xRight, int y, char sym)
         {
             pList = new List<Point>();
-            for (int y = yUp; y <= yDown; y++)
+            for (int x = xLeft; x <= xRight; x++)
             {
                 Point p = new Point(x, y, sym);
                 pList.Add(p);
             }
         }
-
-        public void Draw()
-        {
-            foreach (Point p in pList)
-            {
-                p.Draw();
-            }
-        }
-        
     }
 }
